Movement of General Units looks better now
The reason for that is that Units now will not come to an instant stop,
but decelerate when reaching their target (platform). Still, this needs
further improvement since the wouldn't need to slow down if their actual
goal is a different platform altogether, so having a velocity-vector
that is weighted would be a better looking choice.